Pull out ICNS advection class in preparation for inserting MacProj
