+++ amr-wind/equation_systems/icns/icns.cpp
@@ -1,5 +1,6 @@
 #include "amr-wind/equation_systems/icns/icns.H"
 #include "amr-wind/equation_systems/icns/icns_ops.H"
+#include "amr-wind/equation_systems/icns/icns_advection.H"
 #include "amr-wind/equation_systems/icns/icns_diffusion.H"
 #include "amr-wind/equation_systems/icns/icns_bcop.H"
